Rendered all events from attribute panel  to configure panel
Rendered all events from attribute panel  to configure panel
