+++ src/attr-panel/attr-grid/AttrGridComp.jsx
@@ -23,53 +23,8 @@ const AttrGrid = (props) => {
       element.ref.current.applyGridOptions(); //calls HDGrid.applyGridOptions()
   };
 
-  const ds = ["API-1", "API-2"];
-  const dsSQL = meta.sqlList;
-
-  const handleSQLDataSourceChange = (e) => {
-    handleAttributeChange(e);
-    console.log(e.value);
-    const selectedDs = e.value;
-    selectedDs.extraParam.controlIds.push(element.id);
-    dataConnector.handleDatasourceChange(element);
-  };
-
   return (
     <>
-      <label>
-        <b>Grid Attributes</b>
-        <i>(API as datasource is the first implementation)</i>
-      </label>
-      <div className="field col-12">
-        <label htmlFor="datasource" className="block">
-          Datasource
-        </label>
-        <Dropdown
-          style={{ width: "100%" }}
-          name="datasource"
-          value={meta.currentElement?.attributes?.datasource}
-          options={ds}
-          onChange={(e) => {
-            handleAttributeChange(e);
-            dataConnector.handleDatasourceChange(element);
-          }}
-          placeholder="Select Datasource"
-        />
-      </div>
-      <div className="field col-12">
-        <label htmlFor="datasource" className="block">
-          SQL Datasource
-        </label>
-        <Dropdown
-          style={{ width: "100%" }}
-          name="sqldatasource"
-          value={meta.currentElement?.attributes?.datasource}
-          options={dsSQL}
-          optionLabel="name"
-          onChange={handleSQLDataSourceChange}
-          placeholder="Select SQL Datasource"
-        />
-      </div>
       <div className="field col-12">
         <label htmlFor="datasource" className="block">
           <b>Response to use</b>(
